+++ src/r3f/components/CameraControls.jsx
@@ -1,8 +1,8 @@
 import { forwardRef, useMemo, useEffect, useContext } from 'react';
 import { useThree, useFrame } from '@react-three/fiber';
-import { EnvironmentControls as EnvironmentControlsImpl } from '../three/controls/EnvironmentControls.js';
-import { GlobeControls as GlobeControlsImpl } from '../three/controls/GlobeControls.js';
-import { useShallowOptions } from './utilities/useOptions.jsx';
+import { EnvironmentControls as EnvironmentControlsImpl } from '../../three/controls/EnvironmentControls.js';
+import { GlobeControls as GlobeControlsImpl } from '../../three/controls/GlobeControls.js';
+import { useShallowOptions } from '../utilities/useOptions.jsx';
 import { TilesRendererContext } from './TilesRenderer.jsx';
 
 const ControlsBaseComponent = forwardRef( function ControlsBaseComponent( props, ref ) {
