+++ src/r3f/index.jsx
@@ -1,4 +1,4 @@
-export * from './TilesRenderer.jsx';
-export * from './TilesAttributionOverlay.jsx';
-export * from './CanvasDOMOverlay.jsx';
-export * from './CameraControls.jsx';
+export * from './components/TilesRenderer.jsx';
+export * from './components/TilesAttributionOverlay.jsx';
+export * from './components/CanvasDOMOverlay.jsx';
+export * from './components/CameraControls.jsx';
